@@ -1,7 +1,5 @@
 package com.abc.foaled.fragments;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
@@ -10,9 +8,14 @@ import android.view.ViewGroup;
 
 import com.abc.foaled.R;
 
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link NotificationSettingsFragment#newInstance} factory method to
+ * create an instance of this fragment.
+ */
 public class NotificationSettingsFragment extends Fragment {
 
-    private OnFragmentInteractionListener mListener;
+
 
     public NotificationSettingsFragment() {
         // Required empty public constructor
@@ -30,42 +33,4 @@ public class NotificationSettingsFragment extends Fragment {
         return inflater.inflate(R.layout.fragment_notification_settings, container, false);
     }
 
-    // TODO: Rename method, update argument and hook method into UI event
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
-        }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        if (context instanceof OnFragmentInteractionListener) {
-            mListener = (OnFragmentInteractionListener) context;
-        } else {
-            throw new RuntimeException(context.toString()
-                    + " must implement OnFragmentInteractionListener");
-        }
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        mListener = null;
-    }
-
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
-    public interface OnFragmentInteractionListener {
-        // TODO: Update argument type and name
-        void onFragmentInteraction(Uri uri);
-    }
 }
